refactor: use functions exposed from utils/sfx.ts
